Moved parameterize function to @sentry/core
