+++ packages/core/src/index.ts
@@ -80,6 +80,7 @@ export { createCheckInEnvelope } from './checkin';
 export { hasTracingEnabled } from './utils/hasTracingEnabled';
 export { isSentryRequestUrl } from './utils/isSentryRequestUrl';
 export { handleCallbackErrors } from './utils/handleCallbackErrors';
+export { parameterize } from './utils/parameterize';
 export {
   spanToTraceHeader,
   spanToJSON,
@@ -108,5 +109,3 @@ const Integrations = INTEGRATIONS;
 
 // eslint-disable-next-line deprecation/deprecation
 export { Integrations };
-
-export { parameterize } from '@sentry/utils';
