+++ packages/utils/src/index.ts
@@ -32,7 +32,6 @@ export * from './url';
 export * from './userIntegrations';
 export * from './cache';
 export * from './eventbuilder';
-export * from './parameterize';
 export * from './anr';
 export * from './lru';
 export * from './buildPolyfills';
